Move to existing BeaconChainMetrics class.
